requirements-dev.txt
-r requirements.txt
pytest==8.3.5
pytest-django==4.11.1
factory-boy==3.3.3
